requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
jsonschema>=4.20.0
huggingface_hub>=0.20.0
requests>=2.31.0
